refactor into custom hook and separate fetching logic from hook
